+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### changed
+
+  - Pagination - Change error message when first and last beyond 10'000
+
 ## [4.2.12] - 2020-03-06
 
 ### Added