+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### changed
 
   - Deployment - When module doesn't exist return 400 instead of 500
+  - NuvlaBox peripheral - enable/disable stream action for video class 
+    peripherals
   - Pagination - Change error message when first and last beyond 10'000
   - Voucher - only map country-name if country code exists